backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Churninator backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "fastapi>=0.117.1",
    "greenlet>=3.2.4",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pydantic-settings>=2.11.0",
    "python-jose>=3.5.0",
    "python-socketio>=5.13.0",
    "redis>=6.4.0",
    "sqlmodel>=0.0.25",
    "uvicorn>=0.37.0",
    "dramatiq[redis]>=1.16.5",
    "playwright",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.10",
    "python-multipart>=0.0.20",
    "bcrypt>=4.1.2",
    "stripe>=12.5.1",
    "google-generativeai>=0.8.5",
    "pillow>=11.3.0",
]

[tool.uv.workspace]
members = ["worker"]
